Исправляет fn createAnnouncementPopup
